updating the crypto info
Editing live in the Github repo as data migration to mew machine broke the site (again) on the machine. The problem with flat-site git repos is that local data sometimes gets corrupted. The USB adaptor to pull the latest off file vault backup is in the other room where the kid is sleeping, so can't go in and get it tonight. Life is complicated.